CBG-5784: relocate shared test helpers out of change_cache_test.go
The seven helpers used by tests across package db were defined in
change_cache_test.go, which is about to move to db/changecachetest. Moving the
file would have taken them with it and broken 177 call sites in the tests that
stay behind.

Relocates the six phase-2 delegations to db/change_cache_helpers_test.go, and
exports the one real helper among them - the tombstoned-entry builder et - as
MakeDeletedTestLogEntry in util_testing.go, since it has consumers on both
sides of the boundary.

Also adds db/changesfeedtest/main_test.go, which was left untracked when the
changes feed tests moved, leaving that package with no TestMain and no bucket
pool.
